02-standardisation/pyproject.toml
[project]
name = "standardisation"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "imagehash>=4.3.2",
    "msgspec>=0.20.0",
    "pillow>=12.0.0",
]
